+++ app/javascript/packs/application.js
@@ -17,6 +17,7 @@ import { library, dom } from '@fortawesome/fontawesome-svg-core';
 import { far } from '@fortawesome/free-regular-svg-icons';
 import { fas } from '@fortawesome/free-solid-svg-icons';
 import 'bootstrap-datepicker';
+import 'bootstrap-select';
 
 // :TODO: for now import all fonts, so ux people can work without problems, optimize later
 library.add(fas, far);